final-project-gopal-kapur-muruganandan-ramakrishnan/requirements.txt
dataclasses==0.6
Django==2.1.3
django-logpipe==0.3.0
djangorestframework==3.9.0
djongo==1.2.30
kafka-python==1.4.4
lru-dict==1.1.6
numpy==1.15.4
pandas==0.23.4
pathlib==1.0.1
py4j==0.10.7
pymongo==3.7.2
pyspark==2.4.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
sqlparse==0.2.4
